Отделение физики. Реализация, чистая, корректная, топлива и профиля топлива

+++ rocket/engine/engine/IEngineComponent.h
@@ -5,6 +5,6 @@ class Engine;
 class IEngineComponent
 {
 	virtual ~IEngineComponent() = default;
-	virtual void update(float dt) = 0;
+	virtual void update(double dt) = 0;
 	virtual [[nodiscard]] std::string getName() const = 0;
 };
